# command-center/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  createdAt: string;
}

export interface TodoInput {
  /** @minLength 1 */
  title: string;
}

export interface TodoUpdate {
  /** @minLength 1 */
  title?: string;
  completed?: boolean;
}

export interface TodoStats {
  total: number;
  completed: number;
  pending: number;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 10 */
  message: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  category: string;
  description: string;
}

export type ListProductsParams = {
search?: string;
sortBy?: ListProductsSortBy;
sortOrder?: ListProductsSortOrder;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  name: 'name',
  price: 'price',
} as const;

export type ListProductsSortOrder = typeof ListProductsSortOrder[keyof typeof ListProductsSortOrder];


export const ListProductsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;