# command-center/lib/api-zod/src/generated/types/todoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TodoStats {
  total: number;
  completed: number;
  pending: number;
}