# playwright-tests/app/lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 10 */
  message: string;
}