# playwright-tests/app/lib/api-zod/src/generated/types/todoUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TodoUpdate {
  /** @minLength 1 */
  title?: string;
  completed?: boolean;
}